Clean up comments from 0c721b7
The last merge included some commented css intended to be toggled in the dev console for easy comparison of possible designs. I prefer not to have these in the main branch, as they'll probably become irrelevant over time and might confuse later.

They can always be recovered from this commit, should they be needed in the future.